backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
python-multipart==0.0.9

# LangChain for Agent Orchestration
langchain==0.1.9
langchain-community==0.0.24

# PDF Parsing
pdfplumber==0.10.4
pypdf==4.0.1

# Google Earth Engine
earthengine-api==0.1.390

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Environment & Config
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database (SQLite for dev)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
redis==5.0.1
fastapi-limiter==0.1.6
celery==5.3.6
